Latest misc fixes (#25)
* Fixed title 'bleeding' over popup
Fixed popup arrows too close from target

* Cleanup, key order and new keys

* Guided Tours plural

* Unused files

* Empty state for steps view when no step has ever been created for a tour

* Removed title and description sorts, incompatible with the way content is stored in the database (language keys)

* Removed title and description sorts, incompatible with the way content is stored in the database (language keys)

* Added empty state
Fixed tour_id filter
Fixed page title not translated

* Fixed page title translation and icon

* Added getEmptyStateQuery to return steps of a tour
Fixed tour_id filter
Removed useless comments

* Empty spaces errors

* Empty spaces errors

+++ administrator/components/com_guidedtours/src/View/Steps/HtmlView.php
@@ -63,6 +63,15 @@ class HtmlView extends BaseHtmlView
     public $activeFilters;
 
     /**
+     * Is this view an Empty State
+     *
+     * @var   boolean
+     *
+     * @since 4.0.0
+     */
+    private $isEmptyState = false;
+
+    /**
      * Display the view.
      *
      * @param   string $tpl The name of the template file to parse; automatically searches through the template paths.
@@ -76,6 +85,10 @@ class HtmlView extends BaseHtmlView
         $this->state         = $this->get('State');
         $this->filterForm    = $this->get('FilterForm');
         $this->activeFilters = $this->get('ActiveFilters');
+
+        if (!\count($this->items) && $this->isEmptyState = $this->get('IsEmptyState')) {
+            $this->setLayout('emptystate');
+        }
 
         // Check for errors.
         if (\count($errors = $this->get('Errors'))) {
@@ -101,8 +114,8 @@ class HtmlView extends BaseHtmlView
 
         $toolbar = Toolbar::getInstance('toolbar');
         $tour_id = $this->state->get('tour_id');
-        $title = GuidedtoursHelper::getTourTitle($this->state->get('tour_id'))->title;
-        ToolbarHelper::title(Text::_('COM_GUIDEDTOURS_STEPS_LIST') . ' : ' . $title);
+        $title = GuidedtoursHelper::getTourTitle($this->state->get('filter.tour_id'))->title;
+        ToolbarHelper::title(Text::sprintf('COM_GUIDEDTOURS_STEPS_LIST', Text::_($title)), 'map-signs');
         $arrow  = Factory::getLanguage()->isRtl() ? 'arrow-right' : 'arrow-left';
 
         ToolbarHelper::link(
@@ -115,7 +128,7 @@ class HtmlView extends BaseHtmlView
             $toolbar->addNew('step.add');
         }
 
-        if ($canDo->get('core.edit.state')) {
+        if (!$this->isEmptyState && $canDo->get('core.edit.state')) {
             $dropdown = $toolbar->dropdownButton('status-group')
                 ->text('JTOOLBAR_CHANGE_STATUS')
                 ->toggleSplit(false)
@@ -136,7 +149,7 @@ class HtmlView extends BaseHtmlView
             }
         }
 
-        if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete')) {
+        if (!$this->isEmptyState && $this->state->get('filter.published') == -2 && $canDo->get('core.delete')) {
             $toolbar->delete('steps.delete')
                 ->text('JTOOLBAR_EMPTY_TRASH')
                 ->message('JGLOBAL_CONFIRM_DELETE')
